+++ CHANGELOG.markdown
@@ -21,4 +21,5 @@ This is the first release of UnCMS, forked from version onespacemedia-cms 4.4.
 * Probably fix the long-standing bug where the page tree gets randomly mangled.
 * The `Video` model in the media app has been retired.
 * `cms.context_processors`, which existed to put settings and the package version into templates, no longer exists. Having settings in the context by default makes it easy to make mistakes which expose secret settings. The package version was only ever there to advertise this software in the admin, but UnCMS no longer cares what your admin looks like.
+* `cms.models.base.BasePge` no longer has `in_navigation`, because this did not make sense for the things `PageBase` is intended for. It is now a field on `Page`.
 * A pile of template overrides were removed.